% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSCEFile.R
\name{saveSCEFile}
\alias{saveSCEFile}
\title{Save SCEmbroider files}
\usage{
saveSCEFile(object, path, assay = "RNA", dims = 1:5, cluster.id = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{path}{Path of directory to save files}

\item{assay}{Name of Assay to make files}

\item{dims}{Which dimensions to use as input features (Only if pca, tsne or umap is missing)}

\item{cluster.id}{If set, save name of the cluster idents with this column instead of default idents}
}
\value{
A Seurat object modified in this function
}
\description{
Save a Seurat Object into SCEmbroider files. These file is used when loading new data in SCEmbroider program. Runs PCA, TSNE, UMAP if do not exists.
}
